+++ database/migrations/2025_08_22_105845_add_address_to_users_table.php
@@ -4,20 +4,19 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     */
+return new class extends Migration {
     public function up(): void {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('address')->nullable()->after('password');
+            if (!Schema::hasColumn('users', 'address')) {
+                $table->string('address')->nullable()->after('password');
+            }
         });
     }
     public function down(): void {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('address');
+            if (Schema::hasColumn('users', 'address')) {
+                $table->dropColumn('address');
+            }
         });
     }
-
 };
